wip CI tests

+++ test/test_ephemeral_devnet.sh
@@ -8,13 +8,16 @@ set -e
 . test/_helpers.sh
 
 echo "fresh bootstrap should give the user1 some funds"
-
 nohup scripts/automate.sh &
 pid=$!
 while [ ! -f ${DEVNET_PATH}/${DEVNET_READY_FLAG} ]; do sleep 5; done
 
-echo $pid
-test/assert_user1_has_funds.sh
+user1_lovelace=$(get_address_biggest_lovelace $(cat private-testnet/address/user1.addr))
+echo $user1_lovelace
+# user should have some lovelace there
+if [ $user1_lovelace == ""]; then
+  exit 1
+
 user1_tx_1=$(get_address_biggest_tx $(cat private-testnet/address/user1.addr))
 kill -9 $pid
 
@@ -24,7 +27,11 @@ pid=$!
 while [ ! -f ${DEVNET_PATH}/${DEVNET_READY_FLAG} ]; do sleep 5; done
 echo $pid
 
-test/assert_user1_has_funds.sh
+user1_lovelace=$(get_address_biggest_lovelace $(cat private-testnet/address/user1.addr))
+echo $user1_lovelace
+# user should have some lovelace there
+if [ $user1_lovelace == ""]; then
+  exit 1
 user1_tx_2=$(get_address_biggest_tx $(cat private-testnet/address/user1.addr))
 kill -9 $pid
